Move api calls to view model - Search
