+++ Homework_5/app/src/main/res/layout/fragment_search.xml
@@ -9,11 +9,13 @@
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar_search"
         android:layout_width="match_parent"
-        android:layout_height="64dp"
+        android:layout_height="wrap_content"
         android:background="@color/color_primary"
         android:elevation="4dp"
+        android:padding="8dp"
+        android:paddingStart="8dp"
+        android:paddingEnd="8dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:titleTextColor="@color/on_color">
@@ -33,16 +35,17 @@
             android:paddingEnd="16dp"
             android:paddingBottom="14dp"
             app:queryHint="@string/search_hint"
-            tools:ignore="TextContrastCheck,VisualLintTextFieldSize"
-            tools:layout_editor_absoluteX="16dp"
-            tools:layout_editor_absoluteY="6dp" />
+            tools:ignore="TextContrastCheck,VisualLintTextFieldSize" />
 
     </androidx.appcompat.widget.Toolbar>
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:padding="8dp"
+        android:orientation="vertical"
+        android:padding="16dp"
+        android:paddingStart="8dp"
+        android:paddingEnd="8dp"
         app:layout_constraintTop_toBottomOf="@id/toolbar_search"
         app:layout_constraintStart_toStartOf="parent">
 
@@ -51,20 +54,17 @@
             style="@style/HeadlineSmallNLv2Left"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            tools:text="@string/recent"
-            android:layout_marginBottom="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            android:text="@string/recent"
+            android:layout_marginStart="8dp"
+            android:layout_marginBottom="8dp" />
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recent_searches_container"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            android:orientation="vertical"
-            app:layout_constraintTop_toBottomOf="@+id/recent_heading"
-            app:layout_constraintStart_toStartOf="parent"/>
+            android:orientation="vertical" />
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+    </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
